Clean up the widget sanitization logic
JIRA: RAIL-4273
